+++ sdk/sdkcore/sdkfos.go
@@ -169,6 +169,11 @@ func (c *FortiSDKClient) ReadGroupsList(params *map[string]interface{}) (mapTmp 
 	HTTPMethod := "POST"
 	path := "/ES/api/flexvm/v1/groups/list"
 	rspKey := "groups"
+	if value, ok := (*params)["accountId"]; ok {
+		if value != 0 {
+			path = "/ES/api/fortiflex/v2/groups/list"
+		}
+	}
 
 	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
 	return
@@ -181,6 +186,11 @@ func (c *FortiSDKClient) ReadGroupsNexttoken(params *map[string]interface{}) (ma
 	HTTPMethod := "POST"
 	path := "/ES/api/flexvm/v1/groups/nexttoken"
 	rspKey := "entitlements"
+	if value, ok := (*params)["accountId"]; ok {
+		if value != 0 {
+			path = "/ES/api/fortiflex/v2/groups/nexttoken"
+		}
+	}
 
 	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
 	return